Move conversion functions, add docstrings
